Refactor data types to Double, improve UI/UX, and add GitHub Pages config
- **Logic Refactor:** Switched `credits` and `adRate` from `Float` to `Double` to fix precision issues with high balances. Implemented `MAX_BALANCE` check.
- **UI Improvements:**
    - Settings: Redesigned User Profile (Row layout), Theme Selector (Chips), and fixed "Ad Rate" input glitch (decoupled state).
    - Splash: Removed artificial delay, improved branding.
    - Home: Fixed credit formatting (no scientific notation).
- **GitHub Pages:** Added `publish_config.yml` workflow for manual deployment of `config/config.json`.
- **Assets:** Updated icons to `Icons.AutoMirrored` and `Icons.Filled` (Material).
- **Quality:** Cleaned up code with Spotless and fixed lint/test issues.

+++ data/src/main/java/com/hexium/nodes/data/AdRepository.kt
@@ -3,9 +3,9 @@ package com.hexium.nodes.data
 interface AdRepository {
     suspend fun getAvailableAds(): Int
     suspend fun getMaxAds(): Int
-    suspend fun getAdRewardRate(): Float
+    suspend fun getAdRewardRate(): Double
     suspend fun getAdExpiryHours(): Int
-    suspend fun getCredits(): Float
+    suspend fun getCredits(): Double
     suspend fun watchAd(): Boolean
     suspend fun getAdHistory(): List<Long>
     suspend fun isLoggedIn(): Boolean
